requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
mediapipe>=0.10.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
roboflow>=1.1.0 